Refactoro out draggable modal
